casts are necessary to build ith gcc 3
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@2896 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Misc/preferences_sdl.h
@@ -61,8 +61,8 @@
 #include <vector>
 #include <cmath>
 
-static inline float log2(int x) { return std::log(x) / std::log(2.0); };
-static inline float exp2(int x) { return std::exp(x + std::log(2.0)); };
+static inline float log2(int x) { return std::log((float) x) / std::log(2.0); };
+static inline float exp2(int x) { return std::exp((float) x + std::log(2.0)); };
 
 #ifdef HAVE_UNISTD_H
 #include <unistd.h>	// for getlogin()
